@@ -1,12 +1,12 @@
 module github.com/yskszk63/renvim
 
-go 1.17
+go 1.24.1
 
 require (
 	github.com/creack/pty v1.1.15
 	github.com/justincormack/go-memfd v0.0.0-20170219213707-6e4af0518993
-	github.com/mattn/go-isatty v0.0.13
-	github.com/neovim/go-client v1.1.7
+	github.com/mattn/go-isatty v0.0.20
+	github.com/neovim/go-client v1.2.1
 )
 
-require golang.org/x/sys v0.0.0-20200116001909-b77594299b42 // indirect
+require golang.org/x/sys v0.31.0 // indirect
